More Code Refactoring
